requirements.txt
asgiref==3.5.0
attrs==21.4.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
coverage==6.3.2
Django==4.0.3
django-crontab==0.7.1
django-daterange-filterspec==2.0.4
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
iniconfig==1.1.1
jarowinkler==1.0.2
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyparsing==3.0.7
pytest==7.1.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
rapidfuzz==2.0.7
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tzdata==2022.1
tzlocal==4.2
